backend/video_generator.py
"""
Video Generation Module
Placeholder for actual video generation implementation
"""
import os
from typing import List

# Note: Scene and AdBrief are defined in main.py
# For now, we'll use dict types to avoid circular imports

class VideoGenerator:
    """Base class for video generation"""
    
    def __init__(self):
        self.api_key = os.getenv("VIDEO_API_KEY")
        self.use_gpu = os.getenv("USE_GPU", "false").lower() == "true"
    
    def generate_video_for_scene(self, scene: Scene, ad_brief: AdBrief) -> str:
        """
        Generate video for a single scene
        
        Returns:
            Path to generated video file
        """
        raise NotImplementedError("Subclass must implement generate_video_for_scene")
    
    def generate_audio_for_scene(self, scene: Scene) -> str:
        """
        Generate TTS audio for a scene
        
        Returns:
            Path to generated audio file
        """
        raise NotImplementedError("Subclass must implement generate_audio_for_scene")
    
    def combine_video_and_audio(self, video_path: str, audio_path: str, output_path: str):
        """Combine video and audio using FFmpeg"""
        import subprocess
        subprocess.run([
            "ffmpeg",
            "-i", video_path,
            "-i", audio_path,
            "-c:v", "copy",
            "-c:a", "aac",
            "-shortest",
            output_path
        ], check=True)


class MockVideoGenerator(VideoGenerator):
    """Mock implementation for testing"""
    
    def generate_video_for_scene(self, scene, ad_brief) -> str:
        # Return placeholder path
        return f"/videos/mock-{scene.description[:20].replace(' ', '-')}.mp4"
    
    def generate_audio_for_scene(self, scene) -> str:
        # Return placeholder path
        return f"/videos/mock-{scene.description[:20].replace(' ', '-')}.wav"


class RunwayMLVideoGenerator(VideoGenerator):
    """Runway ML API implementation"""
    
    def generate_video_for_scene(self, scene: Scene, ad_brief: AdBrief) -> str:
        """
        Generate video using Runway ML API
        
        Requires:
            - RUNWAY_API_KEY in environment
            - runwayml package installed
        """
        try:
            import runway
            from runway import generate
            
            runway.api_key = self.api_key or os.getenv("RUNWAY_API_KEY")
            
            # Generate video
            prompt = f"{scene.description}. Style: {ad_brief.style}. Mood: {ad_brief.mood}"
            result = generate(
                prompt=prompt,
                duration=scene.duration,
                aspect_ratio="16:9"
            )
            
            # Save video
            output_path = f"backend/videos/{ad_brief.productName}-scene-{scene.description[:10]}.mp4"
            result.save(output_path)
            
            return output_path
        except ImportError:
            raise ImportError("Runway ML SDK not installed. Run: pip install runwayml")
        except Exception as e:
            raise Exception(f"Runway ML generation failed: {str(e)}")


def get_video_generator() -> VideoGenerator:
    """Factory function to get appropriate video generator"""
    api_key = os.getenv("RUNWAY_API_KEY") or os.getenv("STABILITY_API_KEY") or os.getenv("VIDEO_API_KEY")
    
    if api_key and "RUNWAY" in os.environ:
        return RunwayMLVideoGenerator()
    else:
        # Default to mock for now
        return MockVideoGenerator()

